betting-research/src/requirements.txt
# Python dependencies for betting prediction market bot

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Kalshi API
kalshi-python>=2.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.0

# Data processing
pandas>=2.1.4
numpy>=1.26.2

# Async
asyncio>=3.4.3

# Websocket
websocket-client>=1.7.0

# Environment variables
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Jupyter notebooks support
jupyter>=1.0.0
ipython>=8.18.1
